@@ -73,7 +73,7 @@
     "gsap": "^3.12.5",
     "loading-points": "^1.0.0",
     "lucide-react": "^0.309.0",
-    "next": "^14.2.3",
+    "next": "^15",
     "next-mdx-remote": "^4.4.1",
     "next-themes": "^0.2.1",
     "react": "^18.2.0",
