@@ -1,6 +1,8 @@
-module.exports = {
+const confs = {
   plugins: {
     tailwindcss: {},
     autoprefixer: {},
   },
 };
+
+export default confs;
